MainActivity: use radio log instead of default logs
Use 'logcat -b radio' to get logs from Rlog.
This is useful because the IMS common framework logs all go to the
radio log, so we have most useful informations in a single logcat
command.
(Unfortunately, uncaught exeptions and some other related system logs
for e.g. sms still go to the default logcat buffer, so this might not
be perfect...)