Ims: give up on extending changeEnabledCapabilities in kotlin
kotlin apparently really cannot do this, even in tree.
This SO answer explains why:
https://stackoverflow.com/questions/49284094/inheritance-from-java-class-with-a-public-method-accepting-a-protected-class-in/49287402#49287402
And also suggests using an intermediate java class just for what we
need to override as we have done, so just cleanup the names and
leftover debug attempt.

While we are here prepare for ims voice support (not there yet)
and cleanup a bit.