Register loop: replace with alarms
The register loop would not be woken up when device sleeps.
Use alarms instead.

For future compatibility to avoid waking up more than necessary we only
wake up once for all Sip connections (for e.g. multiSIM that will allow
multiple)

Note we use setAndAllowWhileIdle, and not setExactAndAllowWhileIdle as
we do not really care about precision: might want to check if we can
have an explicit upper bound if we get >1h wakeups, but in practice
this seems reliable enough for periodic register.