Better filter which track we want from the INVITE sdp
Seen on Three:
Two AMR/8000 tracks, one octet-align=1, the other not.
Currently we have the octet-align=0 (and it costed me to implement it),
so select this one

It is probably possible that some carriers only has octet-align=1
But anyway we'll obviously support all AMR + EVS + PCM + OPUS
configurations in the future right?